refactor: more splitting
